pyproject.toml
[project]
name = "tru-bridge-pulse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-socketio>=5.5.1",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.7",
    "load-dotenv>=0.1.0",
    "websockets>=15.0.1",
]
